chore: extract logic to custom hook
